Add test for PlaybackEffects to/from PlaybackEffectsData conversion
